Rename time_series2.ts to timeSeries2.ts and fix imports
