+++ public/app/core/specs/time_series.test.ts
@@ -1,4 +1,4 @@
-import TimeSeries, { updateLegendValues } from 'app/core/time_series2';
+import TimeSeries, { updateLegendValues } from 'app/core/timeSeries2';
 
 describe('TimeSeries', () => {
   let points, series: any;
