+++ public/app/plugins/panel/graph/specs/graph.test.ts
@@ -4,7 +4,7 @@ import { dateTime, EventBusSrv } from '@grafana/data';
 import { MetricsPanelCtrl } from 'app/angular/panel/metrics_panel_ctrl';
 import { PanelCtrl } from 'app/angular/panel/panel_ctrl';
 import config from 'app/core/config';
-import TimeSeries from 'app/core/time_series2';
+import TimeSeries from 'app/core/timeSeries2';
 
 import { DashboardModel } from '../../../../features/dashboard/state';
 import { graphDirective, GraphElement } from '../graph';
