+++ public/app/plugins/panel/graph/specs/threshold_manager.test.ts
@@ -1,6 +1,6 @@
 import angular from 'angular';
 
-import TimeSeries from 'app/core/time_series2';
+import TimeSeries from 'app/core/timeSeries2';
 
 import { ThresholdManager } from '../threshold_manager';
